Backstory Added. Basic Ship Health System Added
A basic health system has been added.
5 health to start, get hit by a mine and the ship takes 1 damage.
ship is dead if hp is 0 or less.
The start of a health system for the enemy ship has been added, however
the pathing detection sphere counts as sphere for collision with the
current code, thus some of the health system for the enemyship has been
commented out until this issue is fixed.
Bullets now explode on collision with enemy ship to prevent
multi-collision while passing through. However same issue with the
pathing sphere as previously mentioned.