Basic Score System Added
Score is now kept and displayed in ascii mode.
Firing and taking damage reduces score.
Score gets past to highscore if you kill the enemyship (impossible in
current build due to the problems mentioned in the health system commit)
No option to input highscore if you die
Note: at this point is it possible that scores can be negative.