WE HAVE MUSIC/SOUND!!!
notes: MineDrop.wav and Death.wav seem to have half the samplerate than
all the other sounds, and i believe this is the reason they dont play
with Sound().play().

They do however play with Music.play() however this means only 1 stream
of music can be used at a time, so with this implementation we can't
have music in the game state or we miss out on these sound effects.

Also added music functions to the GL_assets.
currently can use Assets.get_music("music1") to work in Arcade_menu due
to a dictionary key error, so i'm just loading the music directly.